+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Updated DSC_Disk to allow volumes to be formatted as Dev Drives
+
 ## [5.1.0] - 2023-02-22
 
 ### Changed